requirements.txt
# Flask and CORS
Flask==3.0.0
flask-cors==4.0.0

# Environment variables
python-dotenv==1.0.0

# OpenAI client (compatible with Groq and other providers)
openai==1.12.0

# LangChain core and community
langchain==0.1.7
langchain-community==0.0.20
langchain-openai==0.0.5

# Pinecone vector database
pinecone-client==3.0.0
langchain-pinecone==0.0.3

# Embeddings providers
cohere==4.47

# Document loaders
pypdf==4.0.0
docx2txt==0.8

# Additional utilities
pydantic==2.6.0
pydantic-settings==2.1.0

# Optional: For production deployment
gunicorn==21.2.0
